pyproject.toml
[project]
name = "complaint-priority-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "tqdm>=4.67.1",
]
